Authentication works
Authentication finally works.

Here's several ways it's NOT implemented:
 - It doesn't use AWS's out of the box Cognito authorizer: that only works if you use less-secure non HttpOnly cookies that can be accessed by client side code
 - It doesn't use a custom lambda authorizer, because I have some lambdas that need to be accessed both by admins and guests, and a custom lambda authorizer can't do that.

Instead, each lambda includes its own authorization logic, and relies on a few helper methods.

Also, the various places that need to return unpublished albums for admins are generally working.

+++ app/src/lambdas/api/deleteAlbumLambda.ts
@@ -1,6 +1,7 @@
 import { APIGatewayProxyEvent, APIGatewayProxyHandler, APIGatewayProxyResult } from 'aws-lambda';
 import { handleHttpExceptions, respondSuccessMessage } from '../../lib/lambda_utils/ApiGatewayResponseHelpers';
 import { HttpMethod, ensureHttpMethod, getAlbumPath } from '../../lib/lambda_utils/ApiGatewayRequestHelpers';
+import { ensureAuthorized } from '../../lib/lambda_utils/AuthorizationHelpers';
 import { deleteAlbum } from '../../lib/gallery/deleteAlbum/deleteAlbum';
 
 /**
@@ -9,6 +10,7 @@ import { deleteAlbum } from '../../lib/gallery/deleteAlbum/deleteAlbum';
 export const handler: APIGatewayProxyHandler = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
     try {
         ensureHttpMethod(event, HttpMethod.DELETE);
+        ensureAuthorized(event);
         const albumPath = getAlbumPath(event);
         await deleteAlbum(albumPath);
         return respondSuccessMessage(event, `Album [${albumPath}] deleted`);
