+++ app/src/lambdas/api/deleteImageLambda.ts
@@ -1,6 +1,7 @@
 import { APIGatewayProxyEvent, APIGatewayProxyHandler, APIGatewayProxyResult } from 'aws-lambda';
 import { handleHttpExceptions, respondSuccessMessage } from '../../lib/lambda_utils/ApiGatewayResponseHelpers';
 import { HttpMethod, ensureHttpMethod, getImagePath } from '../../lib/lambda_utils/ApiGatewayRequestHelpers';
+import { ensureAuthorized } from '../../lib/lambda_utils/AuthorizationHelpers';
 import { deleteImage } from '../../lib/gallery/deleteImage/deleteImage';
 
 /**
@@ -9,6 +10,7 @@ import { deleteImage } from '../../lib/gallery/deleteImage/deleteImage';
 export const handler: APIGatewayProxyHandler = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
     try {
         ensureHttpMethod(event, HttpMethod.DELETE);
+        ensureAuthorized(event);
         const imagePath = getImagePath(event);
         await deleteImage(imagePath);
         return respondSuccessMessage(event, `Image [${imagePath}] deleted`);
